Electronic_Voting_System/requirements.txt
cachetools==3.1.1
dill==0.2.9
Django==1.11.5
django-debug-toolbar==1.8
django-environ==0.4.4
django-widget-tweaks==1.4.1
djangorestframework==3.9.4
enum34==1.1.6
future==0.16.0
futures==3.3.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-auth==1.10.0
google-gax==0.16.0
googleapis-common-protos==1.6.0
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.26.0
httplib2==0.15.0
include==0.2.2
jws==0.1.3
oauth2client==3.0.0
ply==3.8
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0

